changed capitalization of files and other
